Extract throttling logic & structs (for cohesion)
